usuario

+++ sched/app.py
@@ -6,8 +6,11 @@ from flask import request, url_for
 from sched.forms import AppointmentForm
 from sched.models import Appointment
 from sched import filters
-
-
+from flask.ext.login import login_required
+from flask.ext.login import LoginManager, current_user
+from flask.ext.login import login_user, logout_user
+from sched.models import User
+# Use Flask-Login to track current user in Flask's session.
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///sched.db'
 # Use Flask-SQLAlchemy for its engine and session
@@ -17,6 +20,38 @@ app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///sched.db'
 db = SQLAlchemy(app)
 db.Model = Base
 filters.init_app(app)
+
+login_manager = LoginManager()
+login_manager.setup_app(app)
+login_manager.login_view = 'login'
+
+@login_manager.user_loader
+def load_user(user_id):
+    """Flask-Login hook to load a User instance from ID."""
+    return db.session.query(User).get(user_id)
+
+@app.route('/login/', methods=['GET', 'POST'])
+def login():
+    if current_user.is_authenticated():
+        return redirect(url_for('appointment_list'))
+    form = LoginForm(request.form)
+    error = None
+    if request.method == 'POST' and form.validate():
+        email = form.username.data.lower().strip()
+        password = form.password.data.lower().strip()
+        user, authenticated = User.authenticate(db.session.query, email, password)
+        if authenticated:
+            login_user(user)
+            return redirect(url_for('appointment_list'))
+    else:
+        error = 'Incorrect username or password.'
+    return render_template('user/login.html',form=form, error=error)
+
+
+@app.route('/logout/')
+def logout():
+    logout_user()
+    return redirect(url_for('login'))
 
 
 @app.route('/')
@@ -29,6 +64,7 @@ def hello():
 
 
 @app.route('/appointments/')
+@login_required
 def appointment_list():
     """
     >>> appointment_list()
@@ -42,6 +78,7 @@ def appointment_list():
 
 
 @app.route('/appointments/<int:appointment_id>/')
+@login_required
 def appointment_detail(appointment_id):
     """
     >>> appointment_detail(3)
@@ -53,10 +90,13 @@ def appointment_detail(appointment_id):
     if appt is None:
         # Abort with Not Found.
         abort(404)
+    if appt.user_id != current_user.id:
+        abort(403)
     return render_template('appointment/detail.html', appt=appt)
 
 
 @app.route('/appointments/<int:appointment_id>/edit/', methods=['GET', 'POST'])
+@login_required
 def appointment_edit(appointment_id):
     """
     >>> appointment_edit(5)
@@ -67,6 +107,8 @@ def appointment_edit(appointment_id):
     appt = db.session.query(Appointment).get(appointment_id)
     if appt is None:
         abort(404)
+    if appt.user_id != current_user.id:
+        abort(403)
     form = AppointmentForm(request.form, appt)
     if request.method == 'POST' and form.validate():
         form.populate_obj(appt)
@@ -77,10 +119,12 @@ def appointment_edit(appointment_id):
 
 
 @app.route('/appointments/create/', methods=['GET', 'POST'])
+@login_required
 def appointment_create():
     """Provide HTML form to create a new appointment."""
     form = AppointmentForm(request.form)
     if request.method == 'POST' and form.validate():
+        appt = Appointment(user_id=current_user.id)        
         appt = Appointment()
         form.populate_obj(appt)
         db.session.add(appt)
@@ -92,6 +136,7 @@ def appointment_create():
 
 
 @app.route('/appointments/<int:appointment_id>/delete/', methods=['DELETE'])
+@login_required
 def appointment_delete(appointment_id):
     # raise NotImplementedError('DELETE')
     appt = db.session.query(Appointment).get(appointment_id)
@@ -100,6 +145,10 @@ def appointment_delete(appointment_id):
         response = jsonify({'status': 'Not Found'})
         response.status = 404
         return response
+    if appt.user_id != current_user.id:
+        response = jsonify({'status': 'Not match User'})
+        response.status = 403
+        return response       
     db.session.delete(appt)
     db.session.commit()
     return jsonify({'status': 'OK'})
